requirements.txt
aiofiles==0.6.0
aiohttp==3.7.4.post0
aiosqlite==0.17.0
async-generator==1.10
async-timeout==3.0.1
asyncio-extras==1.3.2
asyncpraw==7.3.1
asyncprawcore==2.2.1
attrs==21.2.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
discord-pretty-help==1.3.2
discord-py-slash-command==2.3.1
discord.py==1.7.3
gTTS==2.2.3
idna==2.10
multidict==5.1.0
pycparser==2.20
PyNaCl==1.4.0
python-dotenv==0.18.0
requests==2.25.1
six==1.16.0
typing-extensions==3.10.0.0
update-checker==0.18.0
urllib3==1.26.6
websocket-client==1.1.0
yarl==1.6.3
youtube-dl==2021.6.6
